:truck: Move some test to 'test' directory
